tests/wpt/tests/css/css-masking/parsing/mask-size-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-size with valid values</title>
<link rel="author" title="Sam Weinig" href="mailto:sam@webkit.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-size">
<meta name="assert" content="mask-size supports the full grammar '<bg-size>#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("mask-size", "1px");
test_valid_value("mask-size", "1px auto");
test_valid_value("mask-size", "2% 3%");
test_valid_value("mask-size", "auto");
test_valid_value("mask-size", "auto auto", "auto");
test_valid_value("mask-size", "auto 4%");
test_valid_value("mask-size", "contain");
test_valid_value("mask-size", "cover");
test_valid_value("mask-size", "auto 1px, 2% 3%, contain");
</script>
</body>
</html>
